feat(commons): 🎸 Improve homepage storage panel. Add grid view, thumbnails, and more responsive design
Thumbnails are generated by LW editor using `getDcoumentScreenshot`.
Samples and Templates must have their thumbnails generated manually and
added to the json file. Recent files tthumbnails are generated when the
user open the document and when the document is saved. They are saved in
the localstorage toegether with the list of recent files.